requirements.txt
streamlitsupabasepython-dotenvpandasnumpymatplotlibbcrypt